db_lib/pyproject.toml
[tool.poetry]
name = "db_lib"
version = "0.1.0"
description = ""
authors = ["Ponomarev Andrei"]
packages = [{ include = "db_lib" }]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.7.4"
fastapi-sqlalchemy = "^0.2.1"
psycopg2-binary = "2.9.1"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
